apply previous suggestions
